feat: add DB and Prisma
the data is now stored in the database;
Now the user id is generated on the frontend and stored in the context.
Then his name and this ID is stored on the backend.
A new round is created.
This new user and 4 more mock users are added to the round.
All of them are sent back to the frontend.
If the ID of the returned user is the same as the new one generated in the context,
then its data is displayed as “You”